feat: spitted code for datafetch
