+++ src/components/App/Helper/AskQuestion/index.tsx
@@ -9,7 +9,7 @@ import styled from 'styled-components'
 import { Flex } from '~/components/common/Flex'
 import { Text } from '~/components/common/Text'
 import { useSocket } from '~/hooks/useSockets'
-import { postAskQuestion } from '~/network/fetchGraphData'
+import { postAskQuestion } from '~/network/askQuestion'
 import { useAppStore } from '~/stores/useAppStore'
 import { useDataStore } from '~/stores/useDataStore'
 import { useTeachStore } from '~/stores/useTeachStore'
