+++ src/components/App/Helper/TeachMe/index.tsx
@@ -7,7 +7,7 @@ import * as sphinx from 'sphinx-bridge'
 import styled from 'styled-components'
 import { Flex } from '~/components/common/Flex'
 import { Text } from '~/components/common/Text'
-import { postInstagraph, postTeachMe } from '~/network/fetchGraphData'
+import { postInstagraph, postTeachMe } from '~/network/teach'
 import { useAppStore } from '~/stores/useAppStore'
 import { useDataStore } from '~/stores/useDataStore'
 import { InstagraphResponse, useTeachStore } from '~/stores/useTeachStore'
