+++ src/components/App/SideBar/Trending/index.tsx
@@ -11,7 +11,7 @@ import PlusIcon from '~/components/Icons/PlusIcon'
 import ReloadIcon from '~/components/Icons/ReloadIcon'
 import SentimentDataIcon from '~/components/Icons/SentimentDataIcon'
 import { Flex } from '~/components/common/Flex'
-import { getTrends } from '~/network/fetchGraphData'
+import { getTrends } from '~/network/trends'
 import { useAppStore } from '~/stores/useAppStore'
 import { useDataStore } from '~/stores/useDataStore'
 import { useModal } from '~/stores/useModalStore'
