+++ src/components/AppNew/Helper/SentimentAnalysis/index.tsx
@@ -7,7 +7,7 @@ import { PropagateLoader } from 'react-spinners'
 import styled from 'styled-components'
 import { Button } from '~/components/Button'
 import { Text } from '~/components/common/Text'
-import { getSentimentData } from '~/network/fetchGraphData'
+import { getSentimentData } from '~/network/sentiments'
 import { useAppStore } from '~/stores/useAppStore'
 import { useUserStore } from '~/stores/useUserStore'
 import { colors } from '~/utils/colors'
