+++ src/components/AppNew/Helper/TeachMe/index.tsx
@@ -2,12 +2,14 @@ import { Button } from '@mui/material'
 import { useCallback, useEffect, useRef } from 'react'
 import { PropagateLoader } from 'react-spinners'
 import ReactFlow from 'reactflow'
+import 'reactflow/dist/style.css'
 import { Socket } from 'socket.io-client'
 import * as sphinx from 'sphinx-bridge'
 import styled from 'styled-components'
 import { Flex } from '~/components/common/Flex'
 import { Text } from '~/components/common/Text'
-import { postInstagraph, postTeachMe } from '~/network/fetchGraphData'
+import { useSocket } from '~/hooks/useSockets'
+import { postInstagraph, postTeachMe } from '~/network/teach'
 import { useAppStore } from '~/stores/useAppStore'
 import { useDataStore } from '~/stores/useDataStore'
 import { InstagraphResponse, useTeachStore } from '~/stores/useTeachStore'
@@ -15,8 +17,6 @@ import { useUserStore } from '~/stores/useUserStore'
 import { colors } from '~/utils/colors'
 import { updateBudget } from '~/utils/setBudget'
 import { AskQuestion } from '../AskQuestion'
-import 'reactflow/dist/style.css'
-import { useSocket } from '~/hooks/useSockets'
 
 type ResponseType = {
   tutorial: string
