+++ src/components/AppNew/SideBar/Trending/index.tsx
@@ -8,7 +8,7 @@ import PlayIcon from '~/components/Icons/PlayIcon'
 import PlusIcon from '~/components/Icons/PlusIcon'
 import SentimentDataIcon from '~/components/Icons/SentimentDataIcon'
 import { Flex } from '~/components/common/Flex'
-import { getTrends } from '~/network/fetchGraphData'
+import { getTrends } from '~/network/trends'
 import { useDataStore } from '~/stores/useDataStore'
 import { useModal } from '~/stores/useModalStore'
 import { Trending as TrendingType } from '~/types'
