+++ src/components/Universe/Graph/Cubes/SelectionDataNodes/index.tsx
@@ -2,7 +2,7 @@ import { Segments } from '@react-three/drei'
 import { useFrame } from '@react-three/fiber'
 import { memo, useEffect } from 'react'
 import { useGraphData } from '~/components/DataRetriever'
-import { generateLinksFromNodeData } from '~/network/fetchGraphData'
+import { generateLinksFromNodeData } from '~/network/fetchGraphData/helpers/generateLinksFromNodeData'
 import { useDataStore, useSelectedNode } from '~/stores/useDataStore'
 import { ForceSimulation, runForceSimulation } from '~/transformers/forceSimulation'
 import { GraphData } from '~/types'
