+++ src/network/fetchGraphData/types.ts
@@ -12,19 +12,11 @@ export type TeachData = {
   transcripts: string
 }
 
-/* eslint-disable camelcase */
-export type QuestionData = {
-  search_term: string
-  transcripts: string
-  expertise_level: string
-  question_text: string
-}
-
 export type TopicMapItem = {
   children: string[]
   position: Vector3
 }
 
-export type TopicMap = Map<string, TopicMapItem>
+export type TopicMap = Record<string, TopicMapItem>
 
 export type GuestMap = Map<string, GuestMapChild>
